ready for test

+++ code/inc/w25q_flash_drv.h
@@ -28,6 +28,6 @@ void spi_flash_cs_high();
 uint16_t flash_read_id();
 void flash_write_enable();
 void flash_sector_erase(uint32_t addr);
-void flash_read_data(uint32_t addr,uint8_t *datasto,uint16_t num);
-void flash_write_data(uint32_t addr,uint8_t *datasend,uint16_t num);
+void flash_read_data(uint32_t addr,uint8_t *datasto,int num);
+void flash_write_data(uint32_t addr,uint8_t *datasend,int num);
 #endif
